% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{example1_rosen_grad_hess_check}
\alias{example1_rosen_grad_hess_check}
\title{Example 1: Gradient/Hessian checks for the implemented C++ class of Rosenbrock function}
\usage{
example1_rosen_grad_hess_check()
}
\description{
Gradient/Hessian checks for the implemented C++ class of Rosenbrock function.
}
