% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{example1_rosen_nograd_bfgs}
\alias{example1_rosen_nograd_bfgs}
\title{Example 1: Minimize Rosenbrock function (with numerical gradient) using BFGS}
\usage{
example1_rosen_nograd_bfgs()
}
\description{
Minimize Rosenbrock function (with numerical gradient) using BFGS.
}
