% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{example2_sann_tsp}
\alias{example2_sann_tsp}
\title{Example 1: Solve Travelling Salesman Problem (TSP) using SANN}
\usage{
example2_sann_tsp(distmat, x)
}
\arguments{
\item{distmat}{a distance matrix for storing all pair of locations.}

\item{x}{initial route.}
}
\description{
Solve Travelling Salesman Problem (TSP) using SANN.
}
