% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{roptim}
\alias{roptim}
\alias{roptim-package}
\title{roptim}
\description{
Perform general purpose optimizations in R using C++. A unified wrapper
interface is provided to call C functions of the five optimization algorithms
('Nelder-Mead', 'BFGS', 'CG', 'L-BFGS-B' and 'SANN') underlying optim().
}
\author{
Yi Pan
}
