% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{example1_rosen_other_methods}
\alias{example1_rosen_other_methods}
\title{Example 1: Minimize Rosenbrock function using other methods}
\usage{
example1_rosen_other_methods()
}
\description{
Minimize Rosenbrock function using other methods ("Nelder-Mead"/"CG"/ "L-BFGS-B"/"SANN").
}
