% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{example4_wild_fun}
\alias{example4_wild_fun}
\title{Example 4: Minimize a "wild" function using SANN and BFGS}
\usage{
example4_wild_fun()
}
\description{
Minimize a "wild" function using SANN and BFGS.
}
