% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{roptim}
\alias{roptim}
\alias{roptim-package}
\title{roptim}
\description{
This package provides a unified wrapper interface to the C codes of
the five optimization algorithms ("Nelder-Mead", "BFGS", "CG", "L-BFGS-B"
and "SANN") underlying optim function and enables users performing general
purpose optimization tasks using C++ without reimplementing the optimization
routines.
}
\author{
Yi Pan
}
